% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptcelltype.R
\name{gptcelltype}
\alias{gptcelltype}
\title{Cell type annotation with OpenAI GPT models}
\usage{
gptcelltype(input, tissuename = NULL, model = "gpt-4", topgenenumber = 10)
}
\arguments{
\item{input}{Either the differential gene table returned by Seurat FindAllMarkers() function, or a list of genes.}

\item{tissuename}{Optional input of tissue name.}

\item{model}{A valid GPT-4 or GPT-3.5 model name list on https://platform.openai.com/docs/models. Default is 'gpt-4-32k'.}

\item{topgenenumber}{Number of top differential genes to be used if input is Seurat differential genes.}

\item{openai_key}{The OpenAI key obtained from https://platform.openai.com/account/api-keys The default is NA, which will resulting outputing the prompt itself. If an actual key is provided, then the output will be the celltype annotations from the GPT model specified by the user.}
}
\value{
A vector of cell types when the user provide openai_key or the prompt itself when openai_key = NA (default)
}
\description{
Cell type annotation with OpenAI GPT models
}
\details{
Annotate cell types by OpenAI GPT models in a Seurat pipeline or with a custom gene list. If used in a Seurat pipeline, Seurat FindAllMarkers() function needs to be run first and the differential gene table generated by Seurat will serve as the input. If the input is a custom list of genes, one cell type is identified for each element in the list.  Note: system environment should have variable OPENAI_API_KEY = 'your_api_key' or ''. The OpenAI key can be obtained from https://platform.openai.com/account/api-keys. If '', then output the prompt itself. If an actual key is provided, then the output will be the celltype annotations from the GPT model specified by the user.
}
\examples{
# Gene list as input, OPENAI_API_KEY = NA (default)
Sys.setenv(OPENAI_API_KEY = '')
print(Sys.getenv('OPENAI_API_KEY'))  
gptcelltype(input = list(gs1=c('CD4,CD3D'),gs2='CD14'), tissuename=NULL, model='gpt-4', topgenenumber = 10) 
# Seurat object as input, openai_key = NA (default)
#You have to install Seurat package first to run this example
library(Seurat)
data("pbmc_small")
all.markers <- FindAllMarkers(object = pbmc_small)
Sys.setenv(OPENAI_API_KEY = '')
print(Sys.getenv('OPENAI_API_KEY'))  
prompt <- gptcelltype(all.markers, tissuename='human PBMC')
return(prompt)
}
\author{
Wenpin Hou <wh2526@cumc.columbia.edu>, Zhicheng Ji <zhicheng.ji@duke.edu>
}
